requirements.txt
boto3==1.42.60
botocore==1.42.60
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
jmespath==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
